refactor: Update ResponseReview rating type to Double
Changed the `rating` field in `ResponseReviews.AuthorDetails` from `Int` to `Double` to accommodate decimal ratings.

+++ app/src/main/java/azari/amirhossein/filmora/models/detail/ResponseReviews.kt
@@ -37,7 +37,7 @@ data class ResponseReviews(
             @SerializedName("name")
             val name: String?, // CinemaSerf
             @SerializedName("rating")
-            val rating: Int?, // 6
+            val rating: Double?, // 6
             @SerializedName("username")
             val username: String? // Geronimo1967
         )
